fixes the dryRun

+++ core/system.go
@@ -48,13 +48,10 @@ type ABSystem struct {
 // Supported ABSystemOperation types
 const (
 	// ABSystem operations
-	UPGRADE           = "upgrade"
-	FORCE_UPGRADE     = "force-upgrade"
-	DRY_RUN_UPGRADE   = "dry-run-upgrade"
-	APPLY             = "package-apply"
-	DRY_RUN_APPLY     = "dry-run-package-apply"
-	INITRAMFS         = "initramfs"
-	DRY_RUN_INITRAMFS = "dry-run-initramfs"
+	UPGRADE       = "upgrade"
+	FORCE_UPGRADE = "force-upgrade"
+	APPLY         = "package-apply"
+	INITRAMFS     = "initramfs"
 )
 
 // ABSystem rollback response
@@ -179,7 +176,7 @@ func (s *ABSystem) Rebase(name string, dryRun bool) error {
 //		Applies package changes, and updates the system if an update is available.
 //	INITRAMFS:
 //		Updates the initramfs for the future root, but doesn't update the system.
-func (s *ABSystem) RunOperation(operation ABSystemOperation, deleteBeforeCopy bool) error {
+func (s *ABSystem) RunOperation(operation ABSystemOperation, deleteBeforeCopy bool, dryRun bool) error {
 	PrintVerboseInfo("ABSystem.RunOperation", "starting", operation)
 
 	cq := goodies.NewCleanupQueue()
@@ -230,7 +227,7 @@ func (s *ABSystem) RunOperation(operation ABSystemOperation, deleteBeforeCopy bo
 			return err
 		}
 		if !res {
-			if operation != FORCE_UPGRADE && operation != APPLY && operation != DRY_RUN_APPLY {
+			if operation != FORCE_UPGRADE && operation != APPLY {
 				PrintVerboseErr("ABSystem.RunOperation", 1.1, err)
 				return ErrNoUpdate
 			}
@@ -300,9 +297,7 @@ func (s *ABSystem) RunOperation(operation ABSystemOperation, deleteBeforeCopy bo
 	}
 
 	// Stage 3.1: Delete old images
-	switch operation {
-	case DRY_RUN_UPGRADE, DRY_RUN_APPLY, DRY_RUN_INITRAMFS:
-	default:
+	if !dryRun {
 		err = DeleteAllButLatestImage()
 		if err != nil {
 			PrintVerboseErr("ABSystem.RunOperation", 3.1, err)
@@ -332,7 +327,7 @@ func (s *ABSystem) RunOperation(operation ABSystemOperation, deleteBeforeCopy bo
 
 	var imageName string
 	switch operation {
-	case INITRAMFS, DRY_RUN_INITRAMFS:
+	case INITRAMFS:
 		imageName, err = RetrieveImageForRoot(partPresent.Label)
 		if err != nil {
 			PrintVerboseErr("ABSystem.RunOperation", 3.4, err)
@@ -368,29 +363,31 @@ func (s *ABSystem) RunOperation(operation ABSystemOperation, deleteBeforeCopy bo
 
 	if deleteBeforeCopy || os.Getenv("ABROOT_DELETE_BEFORE_COPY") != "" {
 		PrintVerboseInfo("ABSystemRunOperation", "Deleting future system, this will render the future root temporarily unavailable")
-		err := ClearDirectory(partFuture.Partition.MountPoint, nil)
-		if err != nil {
-			PrintVerboseErr("ABSystem.RunOperation", 4, err)
-			return err
+		if !dryRun {
+			err := ClearDirectory(partFuture.Partition.MountPoint, nil)
+			if err != nil {
+				PrintVerboseErr("ABSystem.RunOperation", 4, err)
+				return err
+			}
 		}
 	}
 
 	abrootTrans := filepath.Join(futureRoot, "abroot-trans")
-	err = OciExportRootFs(
-		"abroot-"+uuid.New().String(),
-		imageRecipe,
-		abrootTrans,
-		futureRoot,
-	)
-	if err != nil {
-		PrintVerboseErr("ABSystem.RunOperation", 4.2, err)
-		return err
+	if !dryRun {
+		err = OciExportRootFs(
+			"abroot-"+uuid.New().String(),
+			imageRecipe,
+			abrootTrans,
+			futureRoot,
+		)
+		if err != nil {
+			PrintVerboseErr("ABSystem.RunOperation", 4.2, err)
+			return err
+		}
 	}
 
 	// Stage 4.1: Delete old images
-	switch operation {
-	case DRY_RUN_UPGRADE, DRY_RUN_APPLY, DRY_RUN_INITRAMFS:
-	default:
+	if !dryRun {
 		err = DeleteAllButLatestImage()
 		if err != nil {
 			PrintVerboseErr("ABSystem.RunOperation", 3.1, err)
@@ -399,10 +396,12 @@ func (s *ABSystem) RunOperation(operation ABSystemOperation, deleteBeforeCopy bo
 	}
 
 	// Stage 4.2: Repair root integrity
-	err = RepairRootIntegrity(futureRoot)
-	if err != nil {
-		PrintVerboseErr("ABSystem.RunOperation", 2.4, err)
-		return err
+	if !dryRun {
+		err = RepairRootIntegrity(futureRoot)
+		if err != nil {
+			PrintVerboseErr("ABSystem.RunOperation", 2.4, err)
+			return err
+		}
 	}
 
 	// Stage 5: Write new abimage.abr and config to future/
@@ -421,10 +420,12 @@ func (s *ABSystem) RunOperation(operation ABSystemOperation, deleteBeforeCopy bo
 		return err
 	}
 
-	err = abimage.WriteTo(futureRoot)
-	if err != nil {
-		PrintVerboseErr("ABSystem.RunOperation", 5.2, err)
-		return err
+	if !dryRun {
+		err = abimage.WriteTo(futureRoot)
+		if err != nil {
+			PrintVerboseErr("ABSystem.RunOperation", 5.2, err)
+			return err
+		}
 	}
 
 	varParent := s.RootM.VarPartition.Parent
@@ -439,16 +440,20 @@ func (s *ABSystem) RunOperation(operation ABSystemOperation, deleteBeforeCopy bo
 		settings.Cnf.PartCryptVar = device
 	}
 
-	err = settings.WriteConfigToFile(filepath.Join(futureRoot, "/usr/share/abroot/abroot.json"))
-	if err != nil {
-		PrintVerboseErr("ABSystem.RunOperation", 5.25, err)
-		return err
+	if !dryRun {
+		err = settings.WriteConfigToFile(filepath.Join(futureRoot, "/usr/share/abroot/abroot.json"))
+		if err != nil {
+			PrintVerboseErr("ABSystem.RunOperation", 5.25, err)
+			return err
+		}
 	}
 
-	err = pkgM.WriteSummaryToRoot(futureRoot)
-	if err != nil {
-		PrintVerboseErr("ABSystem.RunOperation", 5.26, err)
-		return err
+	if !dryRun {
+		err = pkgM.WriteSummaryToRoot(futureRoot)
+		if err != nil {
+			PrintVerboseErr("ABSystem.RunOperation", 5.26, err)
+			return err
+		}
 	}
 
 	if UserStopRequested() {
@@ -463,56 +468,56 @@ func (s *ABSystem) RunOperation(operation ABSystemOperation, deleteBeforeCopy bo
 		return err
 	}
 
-	// Stage (dry): If dry-run, exit here before writing to disk
-	// ------------------------------------------------
-	switch operation {
-	case DRY_RUN_UPGRADE, DRY_RUN_APPLY, DRY_RUN_INITRAMFS:
-		PrintVerboseInfo("ABSystem.RunOperation", "dry-run completed")
-		return nil
-	}
-
 	// Stage 6: Update the bootloader
 	// ------------------------------------------------
 	PrintVerboseSimple("[Stage 6] -------- ABSystemRunOperation")
 
-	chroot, err := NewChroot(
-		futureRoot,
-		partFuture.Partition.Uuid,
-		partFuture.Partition.Device,
-		true,
-		filepath.Join("/var/lib/abroot/etc", partPresent.Label),
-	)
-	if err != nil {
-		PrintVerboseErr("ABSystem.RunOperation", 7.02, err)
-		return err
-	}
-
 	generatedGrubConfigPath := "/boot/grub/grub.cfg"
 
-	grubCommand := fmt.Sprintf(settings.Cnf.UpdateGrubCmd, generatedGrubConfigPath)
-	err = chroot.Execute(grubCommand)
-	if err != nil {
-		PrintVerboseErr("ABSystem.RunOperation", 7.1, err)
-		return err
+	if !dryRun {
+
+		chroot, err := NewChroot(
+			futureRoot,
+			partFuture.Partition.Uuid,
+			partFuture.Partition.Device,
+			true,
+			filepath.Join("/var/lib/abroot/etc", partPresent.Label),
+		)
+		if err != nil {
+			PrintVerboseErr("ABSystem.RunOperation", 7.02, err)
+			return err
+		}
+
+		grubCommand := fmt.Sprintf(settings.Cnf.UpdateGrubCmd, generatedGrubConfigPath)
+		err = chroot.Execute(grubCommand)
+		if err != nil {
+			PrintVerboseErr("ABSystem.RunOperation", 7.1, err)
+			return err
+		}
+
+		err = chroot.Execute(settings.Cnf.UpdateInitramfsCmd) // ensure initramfs is updated
+		if err != nil {
+			PrintVerboseErr("ABSystem.RunOperation", 7.2, err)
+			return err
+		}
+
+		err = chroot.Close()
+		if err != nil {
+			PrintVerboseErr("ABSystem.RunOperation", 7.25, err)
+			return err
+		}
+
 	}
 
-	err = chroot.Execute(settings.Cnf.UpdateInitramfsCmd) // ensure initramfs is updated
-	if err != nil {
-		PrintVerboseErr("ABSystem.RunOperation", 7.2, err)
-		return err
-	}
+	var newKernelVer string
 
-	err = chroot.Close()
-	if err != nil {
-		PrintVerboseErr("ABSystem.RunOperation", 7.25, err)
-		return err
-	}
-
-	newKernelVer := getKernelVersion(filepath.Join(futureRoot, "boot"))
-	if newKernelVer == "" {
-		err := errors.New("could not get kernel version")
-		PrintVerboseErr("ABSystem.RunOperation", 7.26, err)
-		return err
+	if !dryRun {
+		newKernelVer = getKernelVersion(filepath.Join(futureRoot, "boot"))
+		if newKernelVer == "" {
+			err := errors.New("could not get kernel version")
+			PrintVerboseErr("ABSystem.RunOperation", 7.26, err)
+			return err
+		}
 	}
 
 	var rootUuid string
@@ -539,46 +544,49 @@ func (s *ABSystem) RunOperation(operation ABSystemOperation, deleteBeforeCopy bo
 
 		futureInitDir := filepath.Join(initMountpoint, partFuture.Label)
 
-		err = os.RemoveAll(futureInitDir)
-		if err != nil {
-			PrintVerboseWarn("ABSystem.RunOperation", 7.44)
-		}
-		err = os.MkdirAll(futureInitDir, 0o755)
-		if err != nil {
-			PrintVerboseWarn("ABSystem.RunOperation", 7.47, err)
-		}
+		if !dryRun {
+			err = os.RemoveAll(futureInitDir)
+			if err != nil {
+				PrintVerboseWarn("ABSystem.RunOperation", 7.44)
+			}
+			err = os.MkdirAll(futureInitDir, 0o755)
+			if err != nil {
+				PrintVerboseWarn("ABSystem.RunOperation", 7.47, err)
+			}
 
-		err = MoveFile(
-			filepath.Join(futureRoot, "boot", "vmlinuz-"+newKernelVer),
-			filepath.Join(futureInitDir, "vmlinuz-"+newKernelVer),
-		)
-		if err != nil {
-			PrintVerboseErr("ABSystem.RunOperation", 7.5, err)
-			return err
-		}
-		err = MoveFile(
-			filepath.Join(futureRoot, "boot", "initrd.img-"+newKernelVer),
-			filepath.Join(futureInitDir, "initrd.img-"+newKernelVer),
-		)
-		if err != nil {
-			PrintVerboseErr("ABSystem.RunOperation", 7.6, err)
-			return err
-		}
-		err = MoveFile(
-			filepath.Join(futureRoot, "boot", "config-"+newKernelVer),
-			filepath.Join(futureInitDir, "config-"+newKernelVer),
-		)
-		if err != nil {
-			PrintVerboseErr("ABSystem.RunOperation", 7.7, err)
-			return err
-		}
-		err = MoveFile(
-			filepath.Join(futureRoot, "boot", "System.map-"+newKernelVer),
-			filepath.Join(futureInitDir, "System.map-"+newKernelVer),
-		)
-		if err != nil {
-			PrintVerboseErr("ABSystem.RunOperation", 7.8, err)
-			return err
+			err = MoveFile(
+				filepath.Join(futureRoot, "boot", "vmlinuz-"+newKernelVer),
+				filepath.Join(futureInitDir, "vmlinuz-"+newKernelVer),
+			)
+			if err != nil {
+				PrintVerboseErr("ABSystem.RunOperation", 7.5, err)
+				return err
+			}
+			err = MoveFile(
+				filepath.Join(futureRoot, "boot", "initrd.img-"+newKernelVer),
+				filepath.Join(futureInitDir, "initrd.img-"+newKernelVer),
+			)
+			if err != nil {
+				PrintVerboseErr("ABSystem.RunOperation", 7.6, err)
+				return err
+			}
+			err = MoveFile(
+				filepath.Join(futureRoot, "boot", "config-"+newKernelVer),
+				filepath.Join(futureInitDir, "config-"+newKernelVer),
+			)
+			if err != nil {
+				PrintVerboseErr("ABSystem.RunOperation", 7.7, err)
+				return err
+			}
+			err = MoveFile(
+				filepath.Join(futureRoot, "boot", "System.map-"+newKernelVer),
+				filepath.Join(futureInitDir, "System.map-"+newKernelVer),
+			)
+			if err != nil {
+				PrintVerboseErr("ABSystem.RunOperation", 7.8, err)
+				return err
+			}
+
 		}
 
 		rootUuid = initPartition.Uuid
@@ -586,16 +594,18 @@ func (s *ABSystem) RunOperation(operation ABSystemOperation, deleteBeforeCopy bo
 		rootUuid = partFuture.Partition.Uuid
 	}
 
-	err = generateABGrubConf(
-		newKernelVer,
-		futureRoot,
-		rootUuid,
-		partFuture.Label,
-		generatedGrubConfigPath,
-	)
-	if err != nil {
-		PrintVerboseErr("ABSystem.RunOperation", 7.9, err)
-		return err
+	if !dryRun {
+		err = generateABGrubConf(
+			newKernelVer,
+			futureRoot,
+			rootUuid,
+			partFuture.Label,
+			generatedGrubConfigPath,
+		)
+		if err != nil {
+			PrintVerboseErr("ABSystem.RunOperation", 7.9, err)
+			return err
+		}
 	}
 
 	// Stage 7: Sync /etc
@@ -608,13 +618,15 @@ func (s *ABSystem) RunOperation(operation ABSystemOperation, deleteBeforeCopy bo
 
 	// make sure the future etc directories exist, ignoring errors
 	newWorkEtc := fmt.Sprintf("/var/lib/abroot/etc/%s-work", partFuture.Label)
-	os.MkdirAll(newUpperEtc, 0o755)
-	os.MkdirAll(newWorkEtc, 0o755)
+	if !dryRun {
+		os.MkdirAll(newUpperEtc, 0o755)
+		os.MkdirAll(newWorkEtc, 0o755)
 
-	err = EtcBuilder.ExtBuildCommand(oldEtc, futureRoot+"/sysconf", oldUpperEtc, newUpperEtc)
-	if err != nil {
-		PrintVerboseErr("AbSystem.RunOperation", 8, err)
-		return err
+		err = EtcBuilder.ExtBuildCommand(oldEtc, futureRoot+"/sysconf", oldUpperEtc, newUpperEtc)
+		if err != nil {
+			PrintVerboseErr("AbSystem.RunOperation", 8, err)
+			return err
+		}
 	}
 
 	// Stage 8: Mount boot partition
@@ -654,7 +666,7 @@ func (s *ABSystem) RunOperation(operation ABSystemOperation, deleteBeforeCopy bo
 		PrintVerboseErr("ABSystem.RunOperation", 11.1, err)
 		return err
 	}
-	if isPresent {
+	if !dryRun && isPresent {
 		grubCfgCurrent := filepath.Join(tmpBootMount, "grub/grub.cfg")
 		grubCfgFuture := filepath.Join(tmpBootMount, "grub/grub.cfg.future")
 
@@ -693,10 +705,12 @@ func (s *ABSystem) RunOperation(operation ABSystemOperation, deleteBeforeCopy bo
 		}
 	}
 
-	err = s.createFinishedFile()
-	if err != nil {
-		PrintVerboseErr("ABSystem.RunOperation", 11.4, err)
-		return fmt.Errorf("could not write finished file: %w", err)
+	if !dryRun {
+		err = s.createFinishedFile()
+		if err != nil {
+			PrintVerboseErr("ABSystem.RunOperation", 11.4, err)
+			return fmt.Errorf("could not write finished file: %w", err)
+		}
 	}
 
 	PrintVerboseInfo("ABSystem.RunOperation", "upgrade completed")
